refactor workload types
