+++ src/workload_type_test.rs
@@ -1,54 +1,18 @@
-use kube::{client::APIClient, config::Configuration};
-
-use crate::schematic::component::Component;
 use crate::workload_type::*;
 
-use std::collections::BTreeMap;
+struct MockWorkloadType {}
 
-#[test]
-fn test_singleton_kube_name() {
-    let cli = APIClient::new(mock_kube_config());
-
-    let sing = Singleton {
-        name: "de".into(),
-        component_name: "hydrate".into(),
-        instance_name: "squidgy".into(),
-        namespace: "tests".into(),
-        definition: Component {
-            ..Default::default()
-        },
-        params: BTreeMap::new(),
-        client: cli,
-        owner_ref: None,
-    };
-
-    assert_eq!("squidgy", sing.kube_name().as_str());
-}
-
-#[test]
-fn test_replicated_service_kube_name() {
-    let cli = APIClient::new(mock_kube_config());
-
-    let rs = ReplicatedService {
-        name: "de".into(),
-        component_name: "hydrate".into(),
-        instance_name: "dehydrate".into(),
-        namespace: "tests".into(),
-        definition: Component {
-            ..Default::default()
-        },
-        params: BTreeMap::new(),
-        client: cli,
-        owner_ref: None,
-    };
-
-    assert_eq!("dehydrate", rs.kube_name().as_str());  
-}
-
-/// This mock builds a KubeConfig that will not be able to make any requests.
-fn mock_kube_config() -> Configuration {
-    Configuration {
-        base_path: ".".into(),
-        client: reqwest::Client::new(),
+impl WorkloadType for MockWorkloadType {
+    fn add(&self) -> Result<(), failure::Error> {
+        Ok(())
     }
+}
+
+/// This is a canary test to make sure that modify and delete have default implementations.
+#[test]
+fn test_workload_type() {
+    let mwlt = MockWorkloadType{};
+
+    assert!(mwlt.modify().is_err());
+    assert!(mwlt.delete().is_ok());
 }
